Translate namespace indices in values.
If the imported value contains a namespace index (NodeId and
QualifiedName DataTypes), the index from the nodeset file has to be
translated to the actual index of the corresponding namespace in the
server.

Introduced a new struct ServerContext which bundles a pointer to a
UA_Server object and a namespace index mapping table. The ServerContext
object is passed as a user context to the generic functions of the
loader, where previously only the UA_Server was passed.

The mapping table is first populated during nodeset import and later
it's used by the value parser to translate indices.